requirements.txt
Jinja2
python-dotenv
pdflatex
python-docx